views: Improve error handling for Objects that do not exist
If the user gave a URL that had an object that did not exist,
the Django server would crash instead of handling the error.

The behaviour has been updated that, if there is no object
present in the table, an Exception will now be generated.
This exception is now passed onto the user as an Error
message within Django, and the operation that was
requested is rejected by the application.